pyproject.toml
[project]
name = "stat-api"
version = "0.3.0"
description = "Бэкенд для нужд проекта mospolynavigation"
requires-python = ">=3.12"
dependencies = [
    "aiofiles",
    "alembic",
    "fastapi-pagination",
    "fastapi",
    "strawberry-graphql",
    "httpx",
    "python-dotenv",
    "python-multipart",
    "sqlalchemy[asyncio]",
    "uuid",
    "uvicorn",
    "pydantic-settings",
    "apscheduler",
    "pwdlib[argon2]",
    "argparse",
    "transliterate",
    "aiosqlite",
    "asyncpg",
    "pillow",
    "PyJWT",
    "user-agents",
]

[dependency-groups]
dev = [
    "pytest",
    "pytest-cov",
]
lint = [
    "flake8",
]

[tool.pytest.ini_options]
testpaths = ["app/tests"]
norecursedirs = [
    ".git",
    ".venv",
    ".pytest_cache",
    "pytest-cache-files-*",
    "node_modules",
    "web",
    "admin-panel",
]
